fix(build): clear -Werror/-WX warnings breaking CI
Three strict-flag failures on the openadsplus integration branch:

- sqlite_backend.cpp: exec_pragma/escape_pragma_key are used only under OPENADS_WITH_SQLCIPHER; guard their definitions with the same #if so the default (SQLCipher OFF) build does not trip clang -Wunused-function.

- sqlite_connection.cpp: cast the loop index to std::size_t at the vector subscripts to avoid -Wsign-conversion (int -> size_type).

- ace_exports.cpp: wrap the trig_* helpers (return std::string/std::vector/TrigError_) in an extern "C++" block so MSVC /W4 /WX does not raise C4190 for C-linkage functions returning C++-incompatible types.

Behavior-preserving; verified with a strict MSVC x64 build (/W4 /WX) of openads_core (56/56, exit 0).

+++ src/sql_backend/sqlite_connection.cpp
@@ -120,8 +120,9 @@ util::Result<void> load_current_row(sqlite3* db, SqliteTable* tbl) {
         tbl->current_nulls.resize(static_cast<std::size_t>(cols));
         for (int c = 0; c < cols; ++c) {
             bool is_null = false;
-            tbl->current_row[c] = format_sqlite_value(stmt, c, is_null);
-            tbl->current_nulls[c] = is_null;
+            tbl->current_row[static_cast<std::size_t>(c)] =
+                format_sqlite_value(stmt, c, is_null);
+            tbl->current_nulls[static_cast<std::size_t>(c)] = is_null;
         }
         tbl->row_valid = true;
     } else if (rc == SQLITE_DONE) {
